refactor: split timeline_component and timeline_container into modules
